+++ cyoa/db/schema.rb
@@ -36,7 +36,7 @@ ActiveRecord::Schema.define(version: 20150811040200) do
     t.integer  "user_id",                    null: false
     t.string   "subject",                    null: false
     t.date     "due_date",                   null: false
-    t.integer  "estimated_completion_hours", null: false
+    t.float    "estimated_completion_hours", null: false
     t.text     "description",                null: false
     t.datetime "created_at",                 null: false
     t.datetime "updated_at",                 null: false
